Revert "separate the large count test to its own file and config to mitigate internal timeouts"
This reverts commit 65e7dd528089b6e7cc2b98ccc3a38406fa92aad6.
